requirements.txt
Django==5.0
django-crispy-forms==2.1
Pillow==10.1
reportlab==4.0.8
django-cors-headers==4.3
python-dotenv==1.0
